Repackaged behavior:
We need to have a source jar because I have stubs in the solution and I
need to be loaded them via AbstractModule#updatePackagedDescriptor (disgrace in fact).
On the other hand build language does wrong when calculating
dependencies to the build module with sources and classes packed in two
different jars.