MPS-22465 ClassLoader cannot load runtime classes because of the missed language in the MPS repository
Enforcing the special fail-first traversal of the dependencies' graph in
the class loading.
After this change the repository constructed only from the 'core' contributors became inconsistent (because of dependencies from core to the
core.xml.properties, core.xml.sax, etc. in the mps-core plugin).
Now by default an empty environment is not consistent and one has to use
the default configuration (with workbench paths and mps-core plugin).