MPS-23400 invokeAndLater->invokeAndWait for the modules' reading
Firstly, the write action during #initComponent is absolutely legal
(moreover we take the write lock in the MPSCore#initComponent,
MPSPersistence#initComponent).

Secondly, the deadlock which urged us to introduce #invokeAndLater call
was fixed in the platform. (synchronization for ApplicationImpl#getStateStore method,
which is called when acquiring a service)
see MPS-20554, MPS-20566, IDEA-138918