MPS-23812 Fix display of error messages in editor
EditorMessageWithTarget#acceptCell now reuses the logic from
CellFinderUtil to properly determine whether a cell displays a given
property of a node.

Also, show the message even if a property cell for a different property
is the big cell for a node. That is, if an editor for concept C with
properties P and Q only defines a single cell for property P, this cell
will also show errors attached to property Q. Previously such a cell
would have been ignored and the message would not appear.