More intuitive scoping for editor components
Similar to extending concepts an editor component is in scope as soon as a dependency on the declaring language is added.
A checking rule make sure that the extends dependency can be added. Before the user had to first add the the extends dependency to get the ec into scope.
Even using CMD/CTRL + R would not do the thing. It's now the same as for extending concepts. We had many MPS users complainging about this behaviour.