MPS-24006 Error evaluating a simple expression
The problem was caused by a change in BaseLanguage where blocks that are
nested in other compound statements are "unwrapped". The evaluation
engine used to create such a block and annotate it with
ToEvaluateAnnotation for further processing by TransformatorImpl.

The fix is to change TransformatorImpl to process the node passed in
(and its descendants) instead of looking for some related node annotated
with ToEvaluateAnnotation.

The fix makes ToEvaluateAnnotation obsolete so this commit removes it.