MPS-24594 Running tests in-process is slow in case the test outputs anything to stderr/stdout
This performance problem is supposed to appear for two reasons
simultaneously.
In this execution mode (in-process) we use PipedInputStream
and PipedOutputStream which use circular buffer under hood.
Thus one thread is writing via pipe, and another thread is reading via
the IDEA BaseDataReader class. Due to synchronisation specifics in the
pipe stream JDK classes and at the same time waiting algorithm in the
BaseDataReader we experience a noticeable performance downswing.
However we are not able to use BLOCKING mode of the IDEA reader since
it exits whenever it meets the end of the stream (which might happen in
our case).

Thus the key change is to reimplement the blocking reading mechanism (BaseDataReader)
which does not wait for the end of the stream but simply blocks IO
instead.
Also:
an old problem is resolved --
MPS-20353 execution-in-process: test passed is shown incorrectly

Asynchronous text printing is removed by using the DefaultRunListener in
the in-process mode (the same as in the out-of-process test execution)

AND
* Not printing in-process test output to the MPS output since the usual test mode does not do that.
* Removing annoying three dots from test logs
* Removing not used test run listener (it is almost a copy)