MPS-25138 Deadlock in classloading during build
The problem hides in the cross-dependent module classloaders.
A depends on B and vice versa. A tries to load a class which belongs to B
and B tries to load a class which belongs to A.
A invokes load(string, bool, bool) -> loadFromDeps -> B.loadFromSelf
B invokes load(string, bool, bool) -> loadFromDeps -> A.loadFromSelf
Obviously we get a deadlock.

The solution is to make #loadFromSelf method fail-safe and not
synchronized. Instead we always defining a new class but in the case
when the class is already defined in the map we take the previous
recorded value in order to have only one loaded class for each class
fqName.
We have to use a concurrent map instead of a plain map.