MPS-24778 Can't execute tests from ant makefile: 'File accessed outside allowed roots...'
Fix is to add the given module paths to VfsRootAccess's allowed roots.
The trick is that we have to add not the module paths themselves but the
module parent paths. It happens because sometimes MPS tries to reach
sibling module jar, e.g. lang.jar is given and MPS looks for the sources
in the lang-src.jar which is not provided