MPS-25114 New run configuration is created on any time I invoke run action on main method MPS-23790 The wrong test is run: MPS does not consider the model for distinguishing editor tests
Including moduleName together with modelName to resolve models in the
repository when running junit run configurations. The proper way
is to store ids in the persistence however I do not want to break run
configurations compatibility in the minor release.

The way to compare configurations is ugly however I cannot deal with it
because of the poor run configurations language.
Therefore the logic of calculating configuration from a context is duplicated.