Rewriting the XMLModelPersistence. Dropping all the outdated properties (relative path, model name, etc.)
which are held up only by this persistence.
It is the matter of persistence -- I do not want the DefaultModelRoot to
pass a `relative path` or some mysteriously constructed from file `model name` to allow
a persistence client not to store the model name inside data source.
It is still on a implementing side.