Reworking API to ease the migration pain
The default ModelFactory methods implementations allow to work legacy
code as it used to.
Though the mappings string->string in the #load and #create mappings
are no longer filled with some specific keys they used to be filled in
the DefaultModelRoot (like 'relative path' option, 'model name' option on the #load invocation)
hence the code potentially breaks.

On the other side we do not promise to pass any particular keys in the ModelFactory
API thus technically speaking we do not break any contract.
That was the reason I dropped the calculations of these excessive options.
The 'relative path' and 'model name' options are seemed to be added
specifically for the sample XML persistence of our own production.
['Relative path' used to be also in the file-per-root persistence however
the new data source functionality offers enough to implement the old
behavior. Which I did (e.g. MakeDirAModel, JavaToMpsConverter).]