MPS-25505 Cannot start Idea with MPS facet
Reverting the classloader which is used to load the model factory (and
data source rule) services.
The original problem (out-of-process tests fail to start)
is that I used ModelFactory.class.getClassloader()
to load the services however ModelFactory is packed into the
mps-openapi.jar while services are packed into mps-core.jar,
and out-process test execution had been failing on
distrib because test engine starts a new process which executes ClassRunner#main with
the classpath of mps-openapi.jar (but without mps-core.jar [why do we
need mps-openapi.jar is another question]). Thus ModelFactory defining
classloader is the classloader which lacks the mps-core.jar with
services.

After that I changed it to the context classloader which is wrong as we
start idea plugin with context classloader set by IDEA (not reset by
ourselves as we do it in the IDE). This CL lacks mps-core.jar as well.

The correct solution is to load services with the classloader which
loads mps-core.jar which is implmented in this commit.