MPS-25285 Exception in MPS console (started from sources) on attempt to migrate agreement sample
The problem is that refresh was triggered on each file (whilst there are
many files which need refresh in the case of an old project). After
updating to the next version of IDEA platform the TooManyEvents
exception started raining on us. It appeared that the only way we are
able to get out of this problem is to take all the action in the EDT.
Though the aforementioned refresh had been working perfectly just before that we were
forced to look for another solution (since moving all code to the EDT
would totally ruin the EDT responsiveness on the startup).

The easiest solution is to replace a lot of single refreshes with one
enormous refresh. It seems to be good but it does not work.
The example is a sudden creation a stack of directories and a file under
them (ie creating a/b/c/d.mps). Since none of the new directories are in
the vfs snapshot the IDEA tracks only the top directory ('a') but not
the others and not the file.

The correct solution is to mark all the roots dirty in the case of refresh
(VirtualFileUtils#refresh).
See the reassuring comment on the VfsUtil#markDirtyAndRefresh.