MPS-25343 Write access is allowed from write-safe contexts only.
Wrapping the task queue flushing into a transaction.
Obviously we needed to create a write-safe context -- that means
we have only two choices.
Guessing the current modality seems to be impossible since IDEA api does not allow
to track it properly.
Disclosing the ModalityState concept in the MPS api is a possible
solution but it is not justified right now. IDEA api does not allow to
guard our model (smodel it is) with a transaction thus making this API
senseless at least without significant changes.
That leaves us with the transaction.

We see this piece of code as a tool to invoke different tasks
asynchronously. Since we are already bound to the IDEA read/write locks
(why again?) I choose to use the transaction system as such mechanism.
Moreover it is quite natural that we wrap each asynchronous
write/command into a transaction (there are potential changes for the IDEA
model there). As for read being wrapped into a transaction -- that is just
a side-effect of our thread policy -- asynchronous invocations ought
to be executed in the order they were scheduled.