Fix MPS-25698 Import of java sources into baseLanguage models is slow
Main aspect of speeding up is in avoiding executing command, rather doing stuff in update mode.
Replacing a lot of nodes in a command while there are still dynamic references in the model leads to
such command listeners as vcs changes tracker to hang the whole process.
That isn't a good solution and will need to be changed. Proper solution would be to find a way to do
all the work with the models while they're not attached to the repository.